Added and fixed bug case (sent by Kanin) Warning! This introduce a slightly different behavior Please review carefully if this is something you want in kaya or if more rework is needed.
The change in behavior: it's assumed that (under japanese rule), when
it'd be possible for one point to become dame, it will be counted as dame.
Consider this situation:

W W W XX
* * 1 WX
w W W XX

The code will consider the point 1 as filled  because it's assumed
that black can break both dame around W and turns it into a fake eye.

- I - believe this is a far less common case in normal games (where
there aren't usually so many dames like this) than the following
situation:

B W B B B
B W W X B
B W * W B
B B W X B

In this case, the white point should be considered as filled because
if white tries to play one of the point to link up the groups and
make this a real eye, black can play the other one (miai)

There is probably a way to detect this case and isolate it but it's
very similar locally to the case presented above and thus will need
a lot more dev to handle. For now, my suggestion is to go with this
patch and monitor the players' reaction.